32.Automated_MCQ_Generator_Project/Automated_MCQ_Generator-main/requirements.txt
langchain
streamlit
python-dotenv
PyPDF2<3.0
langchain_community
langchain_experimental
langchain-google-genai
pdfreader

#for installation of local packages
-e .